refactor(viewer3d): extract shared viewer3d type definitions
